ETLs/postgres_to_es_refactored/requirements.txt
psycopg2-binary==2.8.6
requests==2.24.0
python-dotenv==0.14.0
backoff==1.10.0